requirements.txt
anyio==3.7.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
outcome==1.2.0
packaging==23.1
protobuf==3.20.3
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
starlette==0.27.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
webdriver-manager==3.8.6
wsproto==1.2.0
